+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,17 @@
 # reef-knot
 
+## 5.5.0
+
+### Minor Changes
+
+- Add Binance Web3 Wallet
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@5.3.0
+  - @reef-knot/wallets-list@2.2.0
+
 ## 5.4.0
 
 ### Minor Changes